sevseg write func correct

+++ inc/encoder.h
@@ -4,7 +4,7 @@
 #include "common.h"
 typedef struct Encoder
 {
-  int8_t position;
+  int16_t position;
   uint8_t resolution;
   int8_t angle;
 } Encoder_t;
